Improving the REPL loading feedback
Hi!

This patch improves user feedback when loading the repl.

I found that the problem is related to how the Javascript Event Loop works. The talk "What the heck is the event loop anyway?" by Philip Roberts was a joy of knowledge!

This patch does not solve the entire problem: The fact that the browser still freezes during the loading and the page does not respond until the repl is loaded makes me crazy.

I will try to do some experiments, and if it works, I send a second patch. Probably some solution using js async feature or Web Workers will do the trick.