migrating repl to a web worker
This patch suggests using a Web Worker for the web REPL. I tried to
modify as little code as possible.

Tested on: Chromium, Firefox, Brave, and Opera.

Included in this patch:

repl-worker.js:

Fengari does not have a straightforward way to deal with Web Workers;
This file evaluates the Worker written in Lua and makes it accessible
in pure js for the Web Worker.

repl-worker.lua:

The Web Worker source code; It's responsible for receiving messages
from the UI and sending the evaluated responses.

repl.fnl:

The Lua pane toggle was removed to make possible to run the full REPL
inside a Web Worker.

shim.lua:

A small adjustment to avoid breaking the test.

init.lua:

Instead of using REPL directly, it communicates via messages with the
Web Worker to avoid freezing the page.

I introduced a "fake loader" that gives the feeling that something is
happening to avoid let the user think that something is broken. This
loader simulates percentage progress for 10 seconds waiting time.

The Lua pane toggle was moved to this file.

Also, there is a loadReplWithoutWebWorker function to keep the shim.lua
working.