migrating repl to a web worker with fallback
Hello,

This patch suggests using a Web Worker for the web REPL. We had an issue
with the last one (https://lists.sr.ht/~technomancy/fennel/patches/12080).

The problem was that we had missing files in the upload process inside
the makefile. This patch solves these problems, and I improved it to
ensure a compatible fallback for browsers that do not support Web
Workers. Also, the shim.lua test still works.

I uploaded a version to make testing easier:
https://gbaptista.com/fennel/ (fennel / 123)

Tested on: Luakit, Nyxt, Chromium, Firefox, Brave, and Opera.

Included in this patch:

makefile:
Adds new files into the upload process.

repl-worker.js:

Fengari does not have a straightforward way to deal with Web Workers;
This file evaluates the Worker written in Lua and makes it accessible
in pure js for the Web Worker.

repl-worker.lua:

The Web Worker source code; It's responsible for receiving messages
from the UI and sending the evaluated responses.

repl.fnl:

The Lua pane toggle was removed to make possible to run the full REPL
inside a Web Worker.

init.lua:

Instead of using REPL directly, it communicates via messages with the
Web Worker to avoid freezing the page.

I introduced a "fake loader" that gives the feeling that something is
happening to avoid let the user think that something is broken. This
loader simulates percentage progress for 10 seconds waiting time.

The Lua pane toggle was moved to this file.

Also, there are compatible modes that make possible to use the REPL in
old browsers or to run tests without any browser.